Some changes after review
