+++ designer/server/src/routes/forms/editor-v2/condition.js
@@ -4,7 +4,6 @@ import Stream from 'node:stream'
 import {
   conditionDataSchemaV2,
   conditionWrapperSchemaV2,
-  getConditionV2,
   slugSchema
 } from '@defra/forms-model'
 import Boom from '@hapi/boom'
@@ -16,17 +15,18 @@ import { sessionNames } from '~/src/common/constants/session-names.js'
 import { addCondition, updateCondition } from '~/src/lib/editor.js'
 import { getValidationErrorsFromSession } from '~/src/lib/error-helper.js'
 import * as forms from '~/src/lib/forms.js'
-import {
-  redirectWithAnchorOrUrl,
-  redirectWithErrors
-} from '~/src/lib/redirect-helper.js'
+import { redirectWithErrors } from '~/src/lib/redirect-helper.js'
 import {
   createConditionSessionState,
-  getConditionSessionState,
-  setConditionSessionState
+  getConditionSessionState
 } from '~/src/lib/session-helper.js'
 import * as viewModel from '~/src/models/forms/editor-v2/condition.js'
-import { editorFormPath } from '~/src/models/links.js'
+import { editorFormPath, editorv2Path } from '~/src/models/links.js'
+import {
+  buildSessionState,
+  processErrorMessages,
+  saveSessionState
+} from '~/src/routes/forms/editor-v2/condition-helper.js'
 import { getForm } from '~/src/routes/forms/editor-v2/helpers.js'
 
 export const ROUTE_PATH_CONDITION = `/library/{slug}/editor-v2/condition/{conditionId}/{stateId?}`
@@ -74,78 +74,6 @@ const conditionWrapperSchema = conditionWrapperSchemaV2.keys({
   })
 })
 
-/**
- * @param {Yar} yar
- * @param {string} stateId
- * @param {FormDefinition} definition
- * @param {string} conditionId
- * @returns {ConditionSessionState}
- */
-export function buildSessionState(yar, stateId, definition, conditionId) {
-  const state = getConditionSessionState(yar, stateId)
-  if (!state?.id) {
-    const newState = {
-      id: conditionId,
-      stateId,
-      conditionWrapper:
-        conditionId && conditionId !== 'new'
-          ? getConditionV2(definition, conditionId)
-          : /** @type {ConditionWrapperV2} */ ({
-              id: randomUUID(),
-              items: [{ id: randomUUID() }]
-            })
-    }
-    setConditionSessionState(yar, stateId, newState)
-    return newState
-  }
-  return state
-}
-
-/**
- * @param {Yar} yar
- * @param {ConditionWrapperPayload} payload
- * @param {string} stateId
- * @param {Partial<ConditionDataV2>[]} items
- */
-export function saveSessionState(yar, payload, stateId, items) {
-  const { coordinator, displayName } = payload
-  const state = getConditionSessionState(yar, stateId)
-  const newState = {
-    ...state,
-    conditionWrapper: {
-      ...state?.conditionWrapper,
-      items,
-      displayName,
-      coordinator
-    }
-  }
-
-  // @ts-expect-error - dynamic parse so enforcing type is problematic
-  setConditionSessionState(yar, stateId, newState)
-}
-
-/**
- * Filter out unwanted schema errors
- * @param { Error | undefined } error
- */
-export function filterErrors(error) {
-  if (Joi.isError(error)) {
-    error.details = error.details.filter((err) => {
-      return err.type !== 'array.includesRequiredUnknowns'
-    })
-
-    error.details.forEach((err) => {
-      if (err.path.length > 1) {
-        // Must be the a part of a condition item is in error
-        const idx = err.path.at(1)
-        if (typeof idx === 'number') {
-          err.message = `${err.message} for condition ${idx + 1}`
-        }
-      }
-    })
-  }
-}
-
 export default [
   /**
    * @satisfies {ServerRoute<{ Params: { slug: string, conditionId: string, stateId?: string } }>}
@@ -162,15 +90,11 @@ export default [
       // Set up session if not yet exists
       if (!stateId || !getConditionSessionState(yar, stateId)) {
         const newStateId = createConditionSessionState(yar)
-        return redirectWithAnchorOrUrl(
-          h,
-          slug,
-          conditionId,
-          '',
-          newStateId,
-          '',
-          `condition/${conditionId}/${newStateId}`
-        )
+        return h
+          .redirect(
+            editorv2Path(slug, `condition/${conditionId}/${newStateId}`)
+          )
+          .code(StatusCodes.SEE_OTHER)
       }
 
       // Get form metadata and definition
@@ -302,7 +226,7 @@ export default [
             saveSessionState(yar, payload, stateId, items)
 
             // Filter out unwanted schema errors
-            filterErrors(error)
+            processErrorMessages(error)
 
             return redirectWithErrors(request, h, error, errorKey)
           }
@@ -324,7 +248,6 @@ export default [
  */
 
 /**
- * @import { ConditionDataV2, ConditionSessionState, ConditionWrapperV2, FormDefinition } from '@defra/forms-model'
+ * @import { ConditionDataV2, ConditionWrapperV2 } from '@defra/forms-model'
  * @import { ServerRoute } from '@hapi/hapi'
- * @import { Yar } from '@hapi/yar'
  */
